streamlit/requirements.txt
joblib==1.5.2
matplotlib==3.10.6
numpy==2.3.3
pandas==2.3.2
python-dotenv==1.1.1
scikit-learn==1.7.2
scipy==1.16.1
seaborn==0.13.2
streamlit==1.50.0
torch==2.8.0
torchvision==0.23.0
xgboost==3.0.5
streamlit
pandas
matplotlib
fpdf2